backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0
akshare>=1.10.0
yfinance>=0.2.0
redis>=4.5.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
pandas>=2.0.0
numpy>=1.24.0
pytest>=7.4.0
httpx>=0.24.0